requirements.txt
# Web App
streamlit==1.31.0
plotly==5.18.0

# Data Analysis
pandas==2.2.0
scikit-learn==1.4.0

# Data Collection
requests==2.31.0
beautifulsoup4==4.12.3
pytrends==4.9.2
google-api-python-client==2.120.0
selenium==4.18.1
instaloader==4.11

# Environment & Database
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10

